+++ src/main/java/demo/repository/PostRepository.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface PostRepository {
     public Post find(long id);
     public List<Post> findPage();
-    public long count();
-    public long add(Post post);
+    public int count();
+    public int add(Post post);
     public void modify(Post post);
     public void delete(long id);
 }
